new_admin_panel_sprint_3/01_etl/requirements.txt
flake8==4.0
psycopg2-binary==2.9
elasticsearch==8.1.3
pydantic==1.9.0
python-dotenv==0.20